+++ packages/waku/src/lib/vite-plugin/rsc-analyze-plugin.ts
@@ -1,12 +1,24 @@
 import path from 'node:path';
+import fs from 'node:fs';
 import type { Plugin } from 'vite';
 import * as swc from '@swc/core';
 
 export function rscAnalyzePlugin(
-  clientEntryCallback: (id: string) => void,
-  serverEntryCallback: (id: string) => void,
-  dependecyCallback: (id: string, depId: string) => void,
+  commonFileSet: Set<string>,
+  clientEntryFileSet: Set<string>,
+  serverEntryFileSet: Set<string>,
 ): Plugin {
+  const dependencyMap = new Map<string, Set<string>>();
+  const clientEntryCallback = (id: string) => clientEntryFileSet.add(id);
+  const serverEntryCallback = (id: string) => serverEntryFileSet.add(id);
+  const dependencyCallback = (id: string, depId: string) => {
+    let depSet = dependencyMap.get(id);
+    if (!depSet) {
+      depSet = new Set();
+      dependencyMap.set(id, depSet);
+    }
+    depSet.add(depId);
+  };
   return {
     name: 'rsc-analyze-plugin',
     async transform(code, id) {
@@ -30,12 +42,59 @@ export function rscAnalyzePlugin(
           if (item.type === 'ImportDeclaration') {
             const resolvedId = await this.resolve(item.source.value, id);
             if (resolvedId) {
-              dependecyCallback(id, resolvedId.id);
+              dependencyCallback(id, resolvedId.id);
             }
           }
         }
       }
       return code;
+    },
+    generateBundle(_options, bundle) {
+      // TODO the logic in this function should probably be redesigned.
+      const outputIds = Object.values(bundle).flatMap((item) =>
+        'facadeModuleId' in item && item.facadeModuleId
+          ? [item.facadeModuleId]
+          : [],
+      );
+      const possibleCommonFileMap = new Map<
+        string,
+        { fromClient?: true; notFromClient?: true }
+      >();
+      const seen = new Set<string>();
+      const loop = (id: string, isClient: boolean) => {
+        if (seen.has(id)) {
+          return;
+        }
+        seen.add(id);
+        isClient = isClient || clientEntryFileSet.has(id);
+        dependencyMap.get(id)?.forEach((depId) => {
+          if (!fs.existsSync(depId)) {
+            // HACK is there a better way?
+            return;
+          }
+          let value = possibleCommonFileMap.get(depId);
+          if (!value) {
+            value = {};
+            possibleCommonFileMap.set(depId, value);
+          }
+          if (isClient) {
+            value.fromClient = true;
+          } else {
+            value.notFromClient = true;
+          }
+          loop(depId, isClient);
+        });
+      };
+      outputIds.forEach((id) => loop(id, false));
+      clientEntryFileSet.forEach((id) => loop(id, true));
+      serverEntryFileSet.forEach((id) => loop(id, false));
+      for (const [id, val] of possibleCommonFileMap) {
+        if (val.fromClient && val.notFromClient) {
+          commonFileSet.add(id);
+        }
+      }
+      clientEntryFileSet.forEach((id) => commonFileSet.delete(id));
+      serverEntryFileSet.forEach((id) => commonFileSet.delete(id));
     },
   };
 }
